requirements.txt
gradio
python-docx
pdf2docx
fpdf2